Regular Commit

+++ server/tools/tools.virsh.js
@@ -131,12 +131,16 @@ function resizeNBDFileSystem(cSrv){
 
 function releaseNBD(cSrv){
 	var deferred = Q.defer();
-	getNBDPID(cSrv).then(function(result){
-		tools.runLocalCommand("sudo kill -SIGTERM "+cSrv.NBDPID).then(function(result){
-			console.log(result);
-			deferred.resolve(cSrv);
+	if(cSrv.NBDPID != 0){
+		getNBDPID(cSrv).then(function(result){
+			tools.runLocalCommand("sudo kill -SIGTERM "+cSrv.NBDPID).then(function(result){
+				console.log(result);
+				deferred.resolve(cSrv);
+			}).fail(deferred.reject);
 		}).fail(deferred.reject);
-	}).fail(deferred.reject);
+	} else {
+		deferred.resolve(cSrv);
+	}
 	return deferred.promise;
 }
 
@@ -144,6 +148,7 @@ function getNBDPID(cSrv){
 	var deferred = Q.defer();
 	tools.runLocalCommand("ps aux | grep qemu-nbd").then(function(result) {
 		result = result.trim().split('\n');
+		cSrv.NBDPID = 0;
 		for(var i = 0; i < result.length; i++){
 			if(result[i].indexOf(cSrv.id+'.qcow2') >= 0){
 				cSrv.NBDPID = tools.splitBySpace(result[i])[1];
